migrate mcp manager to v2 1/n
