+++ sgl-router/src/mcp/mod.rs
@@ -7,18 +7,15 @@
 // - Resources: File/data access with subscription support
 // - OAuth: Secure authentication for remote servers
 
-pub mod client_manager;
+pub mod manager;
 pub mod config;
 pub mod connection_pool;
 pub mod error;
 pub mod inventory;
-pub mod manager;
-mod manager_v2;
 pub mod oauth;
 pub mod proxy;
 
 // Re-export the main types for convenience
-pub use client_manager::McpClientManager;
 pub use config::{
     InventoryConfig, McpConfig, McpPoolConfig, McpProxyConfig, McpServerConfig, McpTransport,
     PromptInfo, ResourceInfo, ToolInfo, WarmupServer,
